Update rule definitions
