+++ rct229/rulesets/ashrae9012019/section5/section5rule17.py
@@ -1,30 +1,16 @@
-from rct229.rule_engine.rule_base import RuleDefinitionBase
+from rct229.rule_engine.partial_rule_definition import PartialRuleDefinition
 from rct229.rule_engine.rule_list_indexed_base import RuleDefinitionListIndexedBase
 from rct229.rule_engine.ruleset_model_factory import produce_ruleset_model_instance
 from rct229.rulesets.ashrae9012019 import BASELINE_0
 from rct229.schema.schema_enums import SchemaEnums
-from rct229.rulesets.ashrae9012019.ruleset_functions.get_area_type_window_wall_area_dict import (
-    get_area_type_window_wall_area_dict,
-)
-from rct229.rulesets.ashrae9012019.ruleset_functions.get_opaque_surface_type import (
-    OpaqueSurfaceType as OST,
-)
-from rct229.rulesets.ashrae9012019.ruleset_functions.get_opaque_surface_type import (
-    get_opaque_surface_type,
-)
-from rct229.rulesets.ashrae9012019.ruleset_functions.get_surface_conditioning_category_dict import (
-    SurfaceConditioningCategory as SCC,
-)
-from rct229.rulesets.ashrae9012019.ruleset_functions.get_surface_conditioning_category_dict import (
-    get_surface_conditioning_category_dict,
-)
-from rct229.utils.assertions import getattr_
 from rct229.utils.jsonpath_utils import find_all
-from rct229.utils.pint_utils import ZERO, CalcQ
-from rct229.utils.std_comparisons import std_equal
 
-DOOR = SchemaEnums.schema_enums["SubsurfaceClassificationOptions"].DOOR
-FAIL_MSG = "The vertical fenestration is not distributed across baseline opaque surfaces in the same proportion as in the proposed design. Verify if envelope is existing or altered and can be excluded from this check."
+GENERAL_STATUS = SchemaEnums.schema_enums["SpaceStatusOptions2019ASHRAE901"]
+
+APPLICABLE_GENERAL_STATUS = [
+    GENERAL_STATUS.EXISTING,
+    GENERAL_STATUS.ALTERED,
+]
 
 
 class Section5Rule17(RuleDefinitionListIndexedBase):
@@ -33,143 +19,59 @@ class Section5Rule17(RuleDefinitionListIndexedBase):
     def __init__(self):
         super(Section5Rule17, self).__init__(
             rmrs_used=produce_ruleset_model_instance(
-                USER=False, BASELINE_0=True, PROPOSED=True
+                USER=False, BASELINE_0=True, PROPOSED=False
             ),
-            required_fields={
-                "$": ["weather"],
-                "weather": ["climate_zone"],
-            },
-            each_rule=Section5Rule17.BuildingRule(),
-            index_rmr=BASELINE_0,
             id="5-17",
-            description="The vertical fenestration shall be distributed on each face of the building in the same proportion as in the proposed design.",
+            description="The baseline fenestration area for an existing building shall equal the existing "
+            "fenestration area prior to the proposed work.",
             ruleset_section_title="Envelope",
             standard_section="Section G3.1-5(c) Building Envelope Modeling Requirements for the Baseline building",
-            is_primary_rule=True,
-            list_path="ruleset_model_descriptions[0].buildings[*]",
-            data_items={"climate_zone": (BASELINE_0, "weather/climate_zone")},
+            is_primary_rule=False,
+            each_rule=Section5Rule17.ZoneRule(),
+            index_rmr=BASELINE_0,
+            list_path="$.buildings[*].building_segments[*].zones[*]",
         )
 
-    class BuildingRule(RuleDefinitionListIndexedBase):
+    class ZoneRule(PartialRuleDefinition):
         def __init__(self):
-            super(Section5Rule17.BuildingRule, self).__init__(
+            super(Section5Rule17.ZoneRule, self).__init__(
                 rmrs_used=produce_ruleset_model_instance(
-                    USER=False, BASELINE_0=True, PROPOSED=True
+                    USER=False, BASELINE_0=True, PROPOSED=False
                 ),
-                required_fields={},
-                each_rule=Section5Rule17.BuildingRule.AboveGradeWallRule(),
-                index_rmr=BASELINE_0,
-                # list_path and list_filter together determine the list of
-                # above grade walls to be passed to AboveGradeWallRule
-                list_path="$.building_segments[*].zones[*].surfaces[*]",
             )
 
-        def create_data(self, context, data=None):
-            building_b = context.BASELINE_0
-            building_p = context.PROPOSED
-            climate_zone = data["climate_zone"]
+        def is_applicable(self, context, data=None):
+            zone_b = context.BASELINE_0
+            return any(
+                [
+                    status_type_b in APPLICABLE_GENERAL_STATUS
+                    for status_type_b in find_all("$.spaces[*].status_type", zone_b)
+                ]
+            )
 
-            window_wall_areas_dictionary_b = get_area_type_window_wall_area_dict(
-                climate_zone, building_b
-            )
-            window_wall_areas_dictionary_p = get_area_type_window_wall_area_dict(
-                climate_zone, building_p
-            )
+        def get_calc_vals(self, context, data=None):
+            zone_b = context.BASELINE_0
+
+            existing_or_altered_space_list_b = [
+                space_b["id"]
+                for space_b in find_all("$.spaces[*]", zone_b)
+                if space_b.get("status_type") in APPLICABLE_GENERAL_STATUS
+            ]
 
             return {
-                "total_fenestration_area_b": sum(
-                    find_all("$..total_window_area", window_wall_areas_dictionary_b),
-                    ZERO.AREA,
-                ),
-                "total_fenestration_area_p": sum(
-                    find_all("$..total_window_area", window_wall_areas_dictionary_p),
-                    ZERO.AREA,
-                ),
-                "surface_conditioning_category_dict_b": get_surface_conditioning_category_dict(
-                    data["climate_zone"], building_b
-                ),
+                "existing_or_altered_space_list_b": existing_or_altered_space_list_b
             }
 
-        def list_filter(self, context_item, data=None):
-            surface_b = context_item.BASELINE_0
+        def applicability_check(self, context, calc_vals, data):
+            existing_or_altered_space_list_b = calc_vals[
+                "existing_or_altered_space_list_b"
+            ]
 
-            return (
-                get_opaque_surface_type(surface_b) == OST.ABOVE_GRADE_WALL
-                and data["surface_conditioning_category_dict_b"][surface_b["id"]]
-                != SCC.UNREGULATED
-            )
+            return len(existing_or_altered_space_list_b) > 0
 
-        class AboveGradeWallRule(RuleDefinitionBase):
-            def __init__(self):
-                super(Section5Rule17.BuildingRule.AboveGradeWallRule, self).__init__(
-                    rmrs_used=produce_ruleset_model_instance(
-                        USER=False, BASELINE_0=True, PROPOSED=True
-                    ),
-                    required_fields={
-                        "$": ["construction"],
-                        "construction": ["u_factor"],
-                    },
-                    fail_msg=FAIL_MSG,
-                )
+        def get_manual_check_required_msg(self, context, calc_vals=None, data=None):
+            existing_or_altered_space_list_b = calc_vals[
+                "existing_or_altered_space_list_b"
+            ]
 
-            def get_calc_vals(self, context, data=None):
-                above_grade_wall_b = context.BASELINE_0
-                above_grade_wall_p = context.PROPOSED
-
-                def _helper_calc_val(above_grade_wall):
-                    """Helper function for calculating the total fenestration area for an above grade wall"""
-                    return sum(
-                        [
-                            subsurface.get("glazed_area", ZERO.AREA)
-                            + subsurface.get("opaque_area", ZERO.AREA)
-                            for subsurface in find_all(
-                                "subsurfaces[*]", above_grade_wall
-                            )
-                            if (
-                                getattr_(subsurface, "subsurface", "classification")
-                                == DOOR
-                                and (
-                                    subsurface.get("glazed_area", ZERO.AREA)
-                                    > subsurface.get("opaque_area", ZERO.AREA)
-                                )
-                                or (
-                                    getattr_(subsurface, "subsurface", "classification")
-                                    != DOOR
-                                )
-                            )
-                        ],
-                        ZERO.AREA,
-                    )
-
-                return {
-                    "total_fenestration_area_surface_b": CalcQ(
-                        "area", _helper_calc_val(above_grade_wall_b)
-                    ),
-                    "total_fenestration_area_b": CalcQ(
-                        "area", data["total_fenestration_area_b"]
-                    ),
-                    "total_fenestration_area_surface_p": CalcQ(
-                        "area", _helper_calc_val(above_grade_wall_p)
-                    ),
-                    "total_fenestration_area_p": CalcQ(
-                        "area", data["total_fenestration_area_p"]
-                    ),
-                }
-
-            def rule_check(self, context, calc_vals=None, data=None):
-                total_fenestration_area_surface_b = calc_vals[
-                    "total_fenestration_area_surface_b"
-                ]
-                total_fenestration_area_surface_p = calc_vals[
-                    "total_fenestration_area_surface_p"
-                ]
-                total_fenestration_area_b = calc_vals["total_fenestration_area_b"]
-                total_fenestration_area_p = calc_vals["total_fenestration_area_p"]
-
-                return (
-                    total_fenestration_area_b == ZERO.AREA
-                    and total_fenestration_area_p == ZERO.AREA
-                ) or std_equal(
-                    (total_fenestration_area_surface_b / total_fenestration_area_b),
-                    (total_fenestration_area_surface_p / total_fenestration_area_p),
-                )
+            return f"Part or all of spaces listed below is existing or altered. The baseline vertical fenestration area for a existing zone must equal to the fenestration area prior to the proposed scope of work. The baseline fenestration area in zone must be checked manually. ${existing_or_altered_space_list_b}"
