+++ rct229/rulesets/ashrae9012019/section5/section5rule18.py
@@ -1,16 +1,11 @@
 from rct229.rule_engine.partial_rule_definition import PartialRuleDefinition
 from rct229.rule_engine.rule_list_indexed_base import RuleDefinitionListIndexedBase
 from rct229.rule_engine.ruleset_model_factory import produce_ruleset_model_instance
-from rct229.rulesets.ashrae9012019 import BASELINE_0
+from rct229.rulesets.ashrae9012019 import PROPOSED
 from rct229.schema.schema_enums import SchemaEnums
-from rct229.utils.jsonpath_utils import find_all
 
-GENERAL_STATUS = SchemaEnums.schema_enums["SpaceStatusOptions2019ASHRAE901"]
-
-APPLICABLE_GENERAL_STATUS = [
-    GENERAL_STATUS.EXISTING,
-    GENERAL_STATUS.ALTERED,
-]
+SUBSURFACE_DYNAMIC_GLAZING = SchemaEnums.schema_enums["SubsurfaceDynamicGlazingOptions"]
+UNDETERMINED_MSG = "SUBSURFACE INCLUDES MANUALLY CONTROLLED DYNAMIC GLAZING IN THE PROPOSED DESIGN. VERIFY THAT SHGC AND VT WERE MODELED AS THE AVERAGE OF THE MINIMUM AND MAXIMUM SHGC AND VT."
 
 
 class Section5Rule18(RuleDefinitionListIndexedBase):
@@ -19,59 +14,40 @@ class Section5Rule18(RuleDefinitionListIndexedBase):
     def __init__(self):
         super(Section5Rule18, self).__init__(
             rmrs_used=produce_ruleset_model_instance(
-                USER=False, BASELINE_0=True, PROPOSED=False
+                USER=False, BASELINE_0=False, PROPOSED=True
             ),
+            each_rule=Section5Rule18.SubsurfaceRule(),
+            index_rmr=PROPOSED,
             id="5-18",
-            description="The baseline fenestration area for an existing building shall equal the existing "
-            "fenestration area prior to the proposed work.",
+            description="Manually controlled dynamic glazing shall use the average of the minimum and maximum SHGC and VT.",
             ruleset_section_title="Envelope",
-            standard_section="Section G3.1-5(c) Building Envelope Modeling Requirements for the Baseline building",
+            standard_section="Section G3.1-5(a)5 Building Envelope Modeling Requirements for the Proposed design",
             is_primary_rule=False,
-            each_rule=Section5Rule18.ZoneRule(),
-            index_rmr=BASELINE_0,
-            list_path="$.buildings[*].building_segments[*].zones[*]",
+            list_path="ruleset_model_descriptions[0].buildings[*].building_segments[*].zones[*].surfaces[*].subsurfaces[*]",
         )
 
-    class ZoneRule(PartialRuleDefinition):
+    class SubsurfaceRule(PartialRuleDefinition):
         def __init__(self):
-            super(Section5Rule18.ZoneRule, self).__init__(
+            super(Section5Rule18.SubsurfaceRule, self).__init__(
                 rmrs_used=produce_ruleset_model_instance(
-                    USER=False, BASELINE_0=True, PROPOSED=False
+                    USER=False, BASELINE_0=False, PROPOSED=True
                 ),
-            )
-
-        def is_applicable(self, context, data=None):
-            zone_b = context.BASELINE_0
-            return any(
-                [
-                    status_type_b in APPLICABLE_GENERAL_STATUS
-                    for status_type_b in find_all("$.spaces[*].status_type", zone_b)
-                ]
+                required_fields={"$": ["dynamic_glazing_type"]},
+                manual_check_required_msg=UNDETERMINED_MSG,
             )
 
         def get_calc_vals(self, context, data=None):
-            zone_b = context.BASELINE_0
-
-            existing_or_altered_space_list_b = [
-                space_b["id"]
-                for space_b in find_all("$.spaces[*]", zone_b)
-                if space_b.get("status_type") in APPLICABLE_GENERAL_STATUS
-            ]
-
+            subsurface_p = context.PROPOSED
+            subsurface_dynamic_glazing_type_p = subsurface_p["dynamic_glazing_type"]
             return {
-                "existing_or_altered_space_list_b": existing_or_altered_space_list_b
+                "subsurface_dynamic_glazing_type_p": subsurface_dynamic_glazing_type_p
             }
 
         def applicability_check(self, context, calc_vals, data):
-            existing_or_altered_space_list_b = calc_vals[
-                "existing_or_altered_space_list_b"
+            subsurface_dynamic_glazing_type_p = calc_vals[
+                "subsurface_dynamic_glazing_type_p"
             ]
-
-            return len(existing_or_altered_space_list_b) > 0
-
-        def get_manual_check_required_msg(self, context, calc_vals=None, data=None):
-            existing_or_altered_space_list_b = calc_vals[
-                "existing_or_altered_space_list_b"
-            ]
-
-            return f"Part or all of spaces listed below is existing or altered. The baseline vertical fenestration area for a existing zone must equal to the fenestration area prior to the proposed scope of work. The baseline fenestration area in zone must be checked manually. ${existing_or_altered_space_list_b}"
+            return (
+                subsurface_dynamic_glazing_type_p
+                == SUBSURFACE_DYNAMIC_GLAZING.MANUAL_DYNAMIC
+            )
