+++ rct229/rulesets/ashrae9012019/section5/section5rule29.py
@@ -1,11 +1,22 @@
-from rct229.rule_engine.partial_rule_definition import PartialRuleDefinition
+from rct229.rule_engine.rule_base import RuleDefinitionBase
 from rct229.rule_engine.rule_list_indexed_base import RuleDefinitionListIndexedBase
 from rct229.rule_engine.ruleset_model_factory import produce_ruleset_model_instance
-from rct229.rulesets.ashrae9012019 import PROPOSED
-from rct229.schema.schema_enums import SchemaEnums
+from rct229.rulesets.ashrae9012019 import BASELINE_0
+from rct229.rulesets.ashrae9012019.ruleset_functions.get_opaque_surface_type import (
+    OpaqueSurfaceType as OST,
+)
+from rct229.rulesets.ashrae9012019.ruleset_functions.get_opaque_surface_type import (
+    get_opaque_surface_type,
+)
+from rct229.rulesets.ashrae9012019.ruleset_functions.get_surface_conditioning_category_dict import (
+    SurfaceConditioningCategory as SCC,
+)
+from rct229.rulesets.ashrae9012019.ruleset_functions.get_surface_conditioning_category_dict import (
+    get_surface_conditioning_category_dict,
+)
+from rct229.utils.std_comparisons import std_equal
 
-SUBSURFACE_DYNAMIC_GLAZING = SchemaEnums.schema_enums["SubsurfaceDynamicGlazingOptions"]
-UNDETERMINED_MSG = "SUBSURFACE INCLUDES MANUALLY CONTROLLED DYNAMIC GLAZING IN THE PROPOSED DESIGN. VERIFY THAT SHGC AND VT WERE MODELED AS THE AVERAGE OF THE MINIMUM AND MAXIMUM SHGC AND VT."
+TARGET_ABSORPTANCE_THERMAL_EXTERIOR = 0.9
 
 
 class Section5Rule29(RuleDefinitionListIndexedBase):
@@ -14,40 +25,71 @@ class Section5Rule29(RuleDefinitionListIndexedBase):
     def __init__(self):
         super(Section5Rule29, self).__init__(
             rmrs_used=produce_ruleset_model_instance(
-                USER=False, BASELINE_0=False, PROPOSED=True
+                USER=False, BASELINE_0=True, PROPOSED=False
             ),
-            each_rule=Section5Rule29.SubsurfaceRule(),
-            index_rmr=PROPOSED,
+            required_fields={
+                "$": ["weather"],
+                "weather": ["climate_zone"],
+            },
+            each_rule=Section5Rule29.BuildingRule(),
+            index_rmr=BASELINE_0,
             id="5-29",
-            description="Manually controlled dynamic glazing shall use the average of the minimum and maximum SHGC and VT.",
+            description="The baseline roof surfaces shall be modeled using a thermal emittance of 0.9",
             ruleset_section_title="Envelope",
-            standard_section="Section G3.1-5(a)5 Building Envelope Modeling Requirements for the Proposed design",
-            is_primary_rule=False,
-            list_path="ruleset_model_descriptions[0].buildings[*].building_segments[*].zones[*].surfaces[*].subsurfaces[*]",
+            standard_section="Section G3.1-5(f) Building Envelope Modeling Requirements for the Baseline building",
+            is_primary_rule=True,
+            list_path="ruleset_model_descriptions[0].buildings[*]",
+            data_items={"climate_zone": (BASELINE_0, "weather/climate_zone")},
         )
 
-    class SubsurfaceRule(PartialRuleDefinition):
+    class BuildingRule(RuleDefinitionListIndexedBase):
         def __init__(self):
-            super(Section5Rule29.SubsurfaceRule, self).__init__(
+            super(Section5Rule29.BuildingRule, self).__init__(
                 rmrs_used=produce_ruleset_model_instance(
-                    USER=False, BASELINE_0=False, PROPOSED=True
+                    USER=False, BASELINE_0=True, PROPOSED=False
                 ),
-                required_fields={"$": ["dynamic_glazing_type"]},
-                manual_check_required_msg=UNDETERMINED_MSG,
+                each_rule=Section5Rule29.BuildingRule.RoofRule(),
+                index_rmr=BASELINE_0,
+                list_path="$.building_segments[*].zones[*].surfaces[*]",
             )
 
-        def get_calc_vals(self, context, data=None):
-            subsurface_p = context.PROPOSED
-            subsurface_dynamic_glazing_type_p = subsurface_p["dynamic_glazing_type"]
+        def create_data(self, context, data=None):
+            building_b = context.BASELINE_0
             return {
-                "subsurface_dynamic_glazing_type_p": subsurface_dynamic_glazing_type_p
+                "scc_dict_b": get_surface_conditioning_category_dict(
+                    data["climate_zone"], building_b
+                ),
             }
 
-        def applicability_check(self, context, calc_vals, data):
-            subsurface_dynamic_glazing_type_p = calc_vals[
-                "subsurface_dynamic_glazing_type_p"
-            ]
+        def list_filter(self, context_item, data=None):
+            surface_b = context_item.BASELINE_0
             return (
-                subsurface_dynamic_glazing_type_p
-                == SUBSURFACE_DYNAMIC_GLAZING.MANUAL_DYNAMIC
+                get_opaque_surface_type(surface_b) == OST.ROOF
+                and data["scc_dict_b"][surface_b["id"]] != SCC.UNREGULATED
             )
+
+        class RoofRule(RuleDefinitionBase):
+            def __init__(self):
+                super(Section5Rule29.BuildingRule.RoofRule, self).__init__(
+                    rmrs_used=produce_ruleset_model_instance(
+                        USER=False, BASELINE_0=True, PROPOSED=False
+                    ),
+                    required_fields={
+                        "$": ["optical_properties"],
+                        "optical_properties": ["absorptance_thermal_exterior"],
+                    },
+                )
+
+            def get_calc_vals(self, context, data=None):
+                roof_b = context.BASELINE_0
+                return {
+                    "absorptance_thermal_exterior": roof_b["optical_properties"][
+                        "absorptance_thermal_exterior"
+                    ]
+                }
+
+            def rule_check(self, context, calc_vals=None, data=None):
+                return std_equal(
+                    TARGET_ABSORPTANCE_THERMAL_EXTERIOR,
+                    calc_vals["absorptance_thermal_exterior"],
+                )
