+++ rct229/rulesets/ashrae9012019/section5/section5rule31.py
@@ -1,7 +1,7 @@
 from rct229.rule_engine.rule_base import RuleDefinitionBase
 from rct229.rule_engine.rule_list_indexed_base import RuleDefinitionListIndexedBase
 from rct229.rule_engine.ruleset_model_factory import produce_ruleset_model_instance
-from rct229.rulesets.ashrae9012019 import PROPOSED
+from rct229.rulesets.ashrae9012019 import BASELINE_0
 from rct229.rulesets.ashrae9012019.ruleset_functions.get_opaque_surface_type import (
     OpaqueSurfaceType as OST,
 )
@@ -14,17 +14,9 @@ from rct229.rulesets.ashrae9012019.ruleset_functions.get_surface_conditioning_ca
 from rct229.rulesets.ashrae9012019.ruleset_functions.get_surface_conditioning_category_dict import (
     get_surface_conditioning_category_dict,
 )
+from rct229.utils.std_comparisons import std_equal
 
-ABSORPTION_THERMAL_EXTERIOR = 0.9
-UNDETERMINED_MSG = (
-    "Roof surface emittance in the proposed model {absorptance_thermal_exterior} matches that in the "
-    "user model but is not equal to the prescribed default value of 0.9. Verify that the modeled value "
-    "is based on testing in accordance with section 5.5.3.1.1(a). "
-)
-PASS_DIFFERS_MSG = (
-    "Roof thermal emittance is equal to the prescribed default value of 0.9 but differs from the "
-    "thermal emittance in the user model {absorptance_thermal_exterior} "
-)
+TARGET_ABSORPTANCE_SOLAR_EXTERIOR = 0.7
 
 
 class Section5Rule31(RuleDefinitionListIndexedBase):
@@ -33,112 +25,71 @@ class Section5Rule31(RuleDefinitionListIndexedBase):
     def __init__(self):
         super(Section5Rule31, self).__init__(
             rmrs_used=produce_ruleset_model_instance(
-                USER=True, BASELINE_0=False, PROPOSED=True
+                USER=False, BASELINE_0=True, PROPOSED=False
             ),
+            required_fields={
+                "$": ["weather"],
+                "weather": ["climate_zone"],
+            },
             each_rule=Section5Rule31.BuildingRule(),
-            index_rmr=PROPOSED,
+            index_rmr=BASELINE_0,
             id="5-31",
-            description="The proposed roof surfaces shall be modeled using the same thermal emittance as in the user model.",
+            description=" The baseline roof surfaces shall be modeled using a solar reflectance of 0.30",
             ruleset_section_title="Envelope",
-            standard_section="Section G3.1-1(a) Building Envelope Modeling Requirements for the Proposed design",
+            standard_section="Section G3.1-5(g) Building Envelope Modeling Requirements for the Baseline building",
             is_primary_rule=True,
             list_path="ruleset_model_descriptions[0].buildings[*]",
-            data_items={"climate_zone": (PROPOSED, "weather/climate_zone")},
+            data_items={"climate_zone": (BASELINE_0, "weather/climate_zone")},
         )
 
     class BuildingRule(RuleDefinitionListIndexedBase):
         def __init__(self):
             super(Section5Rule31.BuildingRule, self).__init__(
                 rmrs_used=produce_ruleset_model_instance(
-                    USER=True, BASELINE_0=False, PROPOSED=True
+                    USER=False, BASELINE_0=True, PROPOSED=False
                 ),
                 each_rule=Section5Rule31.BuildingRule.RoofRule(),
-                index_rmr=PROPOSED,
+                index_rmr=BASELINE_0,
                 list_path="$.building_segments[*].zones[*].surfaces[*]",
             )
 
         def create_data(self, context, data=None):
-            building_p = context.PROPOSED
+            building_b = context.BASELINE_0
             return {
-                "scc_dict_p": get_surface_conditioning_category_dict(
-                    data["climate_zone"], building_p
+                "scc_dict_b": get_surface_conditioning_category_dict(
+                    data["climate_zone"], building_b
                 ),
             }
 
         def list_filter(self, context_item, data=None):
-            surface_p = context_item.PROPOSED
-            scc = data["scc_dict_p"][surface_p["id"]]
+            surface_b = context_item.BASELINE_0
             return (
-                get_opaque_surface_type(surface_p) == OST.ROOF
-                and scc is not SCC.UNREGULATED
+                get_opaque_surface_type(surface_b) == OST.ROOF
+                and data["scc_dict_b"][surface_b["id"]] != SCC.UNREGULATED
             )
 
         class RoofRule(RuleDefinitionBase):
             def __init__(self):
                 super(Section5Rule31.BuildingRule.RoofRule, self).__init__(
                     rmrs_used=produce_ruleset_model_instance(
-                        USER=True, BASELINE_0=False, PROPOSED=True
+                        USER=False, BASELINE_0=True, PROPOSED=False
                     ),
                     required_fields={
                         "$": ["optical_properties"],
-                        "optical_properties": ["absorptance_thermal_exterior"],
+                        "optical_properties": ["absorptance_solar_exterior"],
                     },
                 )
 
             def get_calc_vals(self, context, data=None):
-                roof_p = context.PROPOSED
-                roof_u = context.USER
-
+                roof_b = context.BASELINE_0
                 return {
-                    "absorptance_thermal_exterior_p": roof_p["optical_properties"][
-                        "absorptance_thermal_exterior"
-                    ],
-                    "absorptance_thermal_exterior_u": roof_u["optical_properties"][
-                        "absorptance_thermal_exterior"
-                    ],
+                    "absorptance_solar_exterior": roof_b["optical_properties"][
+                        "absorptance_solar_exterior"
+                    ]
                 }
 
-            def manual_check_required(self, context, calc_vals=None, data=None):
-                absorptance_thermal_exterior_p = calc_vals[
-                    "absorptance_thermal_exterior_p"
-                ]
-                absorptance_thermal_exterior_u = calc_vals[
-                    "absorptance_thermal_exterior_u"
-                ]
-                return (
-                    absorptance_thermal_exterior_p == absorptance_thermal_exterior_u
-                    and absorptance_thermal_exterior_p != ABSORPTION_THERMAL_EXTERIOR
-                )
-
-            def get_manual_check_required_msg(self, context, calc_vals=None, data=None):
-                absorptance_thermal_exterior_p = calc_vals[
-                    "absorptance_thermal_exterior_p"
-                ]
-                return UNDETERMINED_MSG.format(
-                    absorptance_thermal_exterior=absorptance_thermal_exterior_p
-                )
-
             def rule_check(self, context, calc_vals=None, data=None):
-                return (
-                    calc_vals["absorptance_thermal_exterior_p"]
-                    == ABSORPTION_THERMAL_EXTERIOR
+                return std_equal(
+                    TARGET_ABSORPTANCE_SOLAR_EXTERIOR,
+                    calc_vals["absorptance_solar_exterior"],
                 )
-
-            def get_pass_msg(self, context, calc_vals=None, data=None):
-                """Pre-condition: see rule_check"""
-                absorptance_thermal_exterior_p = calc_vals[
-                    "absorptance_thermal_exterior_p"
-                ]
-                absorptance_thermal_exterior_u = calc_vals[
-                    "absorptance_thermal_exterior_u"
-                ]
-
-                pass_msg = (
-                    PASS_DIFFERS_MSG.format(
-                        absorptance_thermal_exterior=absorptance_thermal_exterior_u
-                    )
-                    if absorptance_thermal_exterior_p != absorptance_thermal_exterior_u
-                    else ""
-                )
-
-                return pass_msg
