+++ rct229/rulesets/ashrae9012019/section5/section5rule32.py
@@ -1,7 +1,7 @@
 from rct229.rule_engine.rule_base import RuleDefinitionBase
 from rct229.rule_engine.rule_list_indexed_base import RuleDefinitionListIndexedBase
 from rct229.rule_engine.ruleset_model_factory import produce_ruleset_model_instance
-from rct229.rulesets.ashrae9012019 import BASELINE_0
+from rct229.rulesets.ashrae9012019 import PROPOSED
 from rct229.rulesets.ashrae9012019.ruleset_functions.get_opaque_surface_type import (
     OpaqueSurfaceType as OST,
 )
@@ -14,9 +14,11 @@ from rct229.rulesets.ashrae9012019.ruleset_functions.get_surface_conditioning_ca
 from rct229.rulesets.ashrae9012019.ruleset_functions.get_surface_conditioning_category_dict import (
     get_surface_conditioning_category_dict,
 )
-from rct229.utils.std_comparisons import std_equal
 
-TARGET_ABSORPTANCE_SOLAR_EXTERIOR = 0.7
+ABSORPTANCE_SOLAR_EXTERIOR = 0.7
+
+UNDETERMINED_MSG = "Roof surface solar reflectance in the proposed model {absorptance_solar_exterior} matches that in the user model but is not equal to the prescribed default value of 0.3. Verify that reflectance was established using aged test data as required in section 5.5.3.1(a)."
+PASS_DIFFERS_MSG_REGULATED = "Roof surface solar reflectance is equal to the prescribed default value of 0.3 but differs from the solar reflectance in the user model {absorptance_solar_exterior}"
 
 
 class Section5Rule32(RuleDefinitionListIndexedBase):
@@ -25,54 +27,52 @@ class Section5Rule32(RuleDefinitionListIndexedBase):
     def __init__(self):
         super(Section5Rule32, self).__init__(
             rmrs_used=produce_ruleset_model_instance(
-                USER=False, BASELINE_0=True, PROPOSED=False
+                USER=True, BASELINE_0=False, PROPOSED=True
             ),
-            required_fields={
-                "$": ["weather"],
-                "weather": ["climate_zone"],
-            },
             each_rule=Section5Rule32.BuildingRule(),
-            index_rmr=BASELINE_0,
-            id="5-42",
-            description=" The baseline roof surfaces shall be modeled using a solar reflectance of 0.30",
+            index_rmr=PROPOSED,
+            id="5-32",
+            description="The proposed roof surfaces shall be modeled using the same solar reflectance as in the user "
+            "model if the aged test data are available, or equal to 0.7 default reflectance",
             ruleset_section_title="Envelope",
-            standard_section="Section G3.1-5(g) Building Envelope Modeling Requirements for the Baseline building",
+            standard_section="Section G3.1-1(a) Building Envelope Modeling Requirements for the Proposed design",
             is_primary_rule=True,
             list_path="ruleset_model_descriptions[0].buildings[*]",
-            data_items={"climate_zone": (BASELINE_0, "weather/climate_zone")},
+            data_items={"climate_zone": (PROPOSED, "weather/climate_zone")},
         )
 
     class BuildingRule(RuleDefinitionListIndexedBase):
         def __init__(self):
             super(Section5Rule32.BuildingRule, self).__init__(
                 rmrs_used=produce_ruleset_model_instance(
-                    USER=False, BASELINE_0=True, PROPOSED=False
+                    USER=True, BASELINE_0=False, PROPOSED=True
                 ),
                 each_rule=Section5Rule32.BuildingRule.RoofRule(),
-                index_rmr=BASELINE_0,
+                index_rmr=PROPOSED,
                 list_path="$.building_segments[*].zones[*].surfaces[*]",
             )
 
         def create_data(self, context, data=None):
-            building_b = context.BASELINE_0
+            building_p = context.PROPOSED
             return {
-                "scc_dict_b": get_surface_conditioning_category_dict(
-                    data["climate_zone"], building_b
+                "scc_dict_p": get_surface_conditioning_category_dict(
+                    data["climate_zone"], building_p
                 ),
             }
 
         def list_filter(self, context_item, data=None):
-            surface_b = context_item.BASELINE_0
+            surface_p = context_item.PROPOSED
+            scc = data["scc_dict_p"][surface_p["id"]]
             return (
-                get_opaque_surface_type(surface_b) == OST.ROOF
-                and data["scc_dict_b"][surface_b["id"]] != SCC.UNREGULATED
+                get_opaque_surface_type(surface_p) == OST.ROOF
+                and scc is not SCC.UNREGULATED
             )
 
         class RoofRule(RuleDefinitionBase):
             def __init__(self):
                 super(Section5Rule32.BuildingRule.RoofRule, self).__init__(
                     rmrs_used=produce_ruleset_model_instance(
-                        USER=False, BASELINE_0=True, PROPOSED=False
+                        USER=True, BASELINE_0=False, PROPOSED=True
                     ),
                     required_fields={
                         "$": ["optical_properties"],
@@ -81,15 +81,51 @@ class Section5Rule32(RuleDefinitionListIndexedBase):
                 )
 
             def get_calc_vals(self, context, data=None):
-                roof_b = context.BASELINE_0
+                roof_p = context.PROPOSED
+                roof_u = context.USER
+                scc_dict_p = data["scc_dict_p"]
+
                 return {
-                    "absorptance_solar_exterior": roof_b["optical_properties"][
+                    "absorptance_solar_exterior_p": roof_p["optical_properties"][
                         "absorptance_solar_exterior"
-                    ]
+                    ],
+                    "absorptance_solar_exterior_u": roof_u["optical_properties"][
+                        "absorptance_solar_exterior"
+                    ],
+                    "surface_conditioning_category_p": scc_dict_p[roof_p["id"]],
                 }
 
-            def rule_check(self, context, calc_vals=None, data=None):
-                return std_equal(
-                    TARGET_ABSORPTANCE_SOLAR_EXTERIOR,
-                    calc_vals["absorptance_solar_exterior"],
+            def manual_check_required(self, context, calc_vals=None, data=None):
+                absorptance_solar_exterior_p = calc_vals["absorptance_solar_exterior_p"]
+                absorptance_solar_exterior_u = calc_vals["absorptance_solar_exterior_u"]
+                return (
+                    absorptance_solar_exterior_p == absorptance_solar_exterior_u
+                    and absorptance_solar_exterior_p != ABSORPTANCE_SOLAR_EXTERIOR
                 )
+
+            def get_manual_check_required_msg(self, context, calc_vals=None, data=None):
+                absorptance_solar_exterior_p = calc_vals["absorptance_solar_exterior_p"]
+                return UNDETERMINED_MSG.format(
+                    absorptance_solar_exterior=absorptance_solar_exterior_p
+                )
+
+            def rule_check(self, context, calc_vals=None, data=None):
+                return (
+                    calc_vals["absorptance_solar_exterior_p"]
+                    == ABSORPTANCE_SOLAR_EXTERIOR
+                )
+
+            def get_pass_msg(self, context, calc_vals=None, data=None):
+                """Pre-condition: see rule_check"""
+                absorptance_solar_exterior_p = calc_vals["absorptance_solar_exterior_p"]
+                absorptance_solar_exterior_u = calc_vals["absorptance_solar_exterior_u"]
+
+                # this condition only applies when P-RMD = 0.7 and P-RMD surface is regulated.
+                pass_msg = (
+                    PASS_DIFFERS_MSG_REGULATED.format(
+                        absorptance_solar_exterior=(1 - absorptance_solar_exterior_u)
+                    )
+                    if absorptance_solar_exterior_p != absorptance_solar_exterior_u
+                    else ""
+                )
+                return pass_msg
