+++ rct229/rulesets/ashrae9012019/section5/section5rule34.py
@@ -1,11 +1,9 @@
 from rct229.rule_engine.rule_base import RuleDefinitionBase
 from rct229.rule_engine.rule_list_indexed_base import RuleDefinitionListIndexedBase
 from rct229.rule_engine.ruleset_model_factory import produce_ruleset_model_instance
-from rct229.rulesets.ashrae9012019 import BASELINE_0
-from rct229.schema.schema_enums import SchemaEnums
+from rct229.rulesets.ashrae9012019 import PROPOSED
 from rct229.utils.jsonpath_utils import find_all
-
-CONSTANT = SchemaEnums.schema_enums["InfiltrationMethodOptions"].CONSTANT
+from rct229.utils.match_lists import match_lists_by_id
 
 
 class Section5Rule34(RuleDefinitionListIndexedBase):
@@ -14,12 +12,12 @@ class Section5Rule34(RuleDefinitionListIndexedBase):
     def __init__(self):
         super(Section5Rule34, self).__init__(
             rmrs_used=produce_ruleset_model_instance(
-                USER=False, BASELINE_0=True, PROPOSED=False
+                USER=False, BASELINE_0=True, PROPOSED=True
             ),
             each_rule=Section5Rule34.BuildingRule(),
-            index_rmr=BASELINE_0,
+            index_rmr=PROPOSED,
             id="5-34",
-            description="The infiltration modeling method in the baseline includes adjustment for weather and building operation.",
+            description="The infiltration shall be modeled using the same methodology and adjustments for weather and building operation in both the proposed design and the baseline building design.",
             ruleset_section_title="Envelope",
             standard_section="Section G3.1-5(b) Building Envelope Modeling Requirements for the Proposed design and Baseline",
             is_primary_rule=True,
@@ -30,25 +28,44 @@ class Section5Rule34(RuleDefinitionListIndexedBase):
         def __init__(self):
             super(Section5Rule34.BuildingRule, self).__init__(
                 rmrs_used=produce_ruleset_model_instance(
-                    USER=False, BASELINE_0=True, PROPOSED=False
+                    USER=False, BASELINE_0=True, PROPOSED=True
                 ),
                 required_fields={
-                    "$.building_segments[*].zones[*].spaces[*].infiltration": [
-                        "modeling_method"
-                    ]
+                    "$.building_segments[*].zones[*]": ["infiltration"],
+                    "$.building_segments[*].zones[*].infiltration": [
+                        "algorithm_name",
+                        "modeling_method",
+                    ],
                 },
             )
 
         def get_calc_vals(self, context, data=None):
-            baseline_infiltration = find_all(
-                "$.building_segments[*].zones[*].infiltration", context.BASELINE_0
+            failing_infiltration_zone_ids = []
+            baseline_zones = find_all(
+                "$.building_segments[*].zones[*]", context.BASELINE_0
             )
-            failing_infiltration_ids = [
-                b_infiltration["id"]
-                for b_infiltration in baseline_infiltration
-                if b_infiltration["modeling_method"] == CONSTANT
-            ]
-            return {"failing_infiltration_ids": failing_infiltration_ids}
+            proposed_zones = find_all(
+                "$.building_segments[*].zones[*]", context.PROPOSED
+            )
+
+            # This assumes that the surfaces all match
+            matched_baseline_zones = match_lists_by_id(proposed_zones, baseline_zones)
+            proposed_baseline_zone_pairs = zip(proposed_zones, matched_baseline_zones)
+            for p_zone, b_zone in proposed_baseline_zone_pairs:
+                # need a method like match object
+                p_zone_infiltration = p_zone["infiltration"]
+                # b_zone could be NONE - add a check.
+                b_zone_infiltration = b_zone["infiltration"]
+
+                if (
+                    p_zone_infiltration["algorithm_name"]
+                    != b_zone_infiltration["algorithm_name"]
+                    or p_zone_infiltration["modeling_method"]
+                    != b_zone_infiltration["modeling_method"]
+                ):
+                    failing_infiltration_zone_ids.append(p_zone["id"])
+
+            return {"failing_infiltration_zone_ids": failing_infiltration_zone_ids}
 
         def rule_check(self, context, calc_vals=None, data=None):
-            return len(calc_vals["failing_infiltration_ids"]) == 0
+            return len(calc_vals["failing_infiltration_zone_ids"]) == 0
